// web/src/enums/command.ts
export enum Command {
  PLAY_ENTRY = "PLAY_ENTRY",
  PAUSE_ENTRY = "PAUSE_ENTRY",
  CREATE_ENTRY = "CREATE_ENTRY",
  START_ENTRY = "START_ENTRY",
  STOP_ENTRY = "STOP_ENTRY",
  RESUME_ENTRY = "RESUME_ENTRY",
  DELETE_ENTRY = "DELETE_ENTRY",
  ECHO_ENTRY = "ECHO_ENTRY",
  RUN_ENTRY = "RUN_ENTRY",
  WALK_ENTRY = "WALK_ENTRY",
  EDIT_ENTRY = "EDIT_ENTRY",
  EXPORT_ENTRY = "EXPORT_ENTRY",
  ENTER_ENTRY = "ENTER_ENTRY",
  WALK_ENTRY_LIST = "WALK_ENTRY_LIST",
  RUN_ENTRY_LIST = "RUN_ENTRY_LIST",
  ENTER_ENTRY_LIST = "ENTER_ENTRY_LIST",
  PLAY_ECHO = "PLAY_ECHO",
  START_ECHO = "START_ECHO",
  PAUSE_ECHO = "PAUSE_ECHO",
  STOP_ECHO = "STOP_ECHO",
  RESUME_ECHO = "RESUME_ECHO",
  PAUSE_PLAYBACK = "PAUSE_PLAYBACK",
  RESUME_PLAYBACK = "RESUME_PLAYBACK",
  STOP_PLAYBACK = "STOP_PLAYBACK",
  SKIP_PLAYBACK_BACKWARD = "SKIP_PLAYBACK_BACKWARD",
  SKIP_PLAYBACK_FORWARD = "SKIP_PLAYBACK_FORWARD",
  ACTIVATE_PUNCTUATION = "ACTIVATE_PUNCTUATION",
  DEACTIVATE_PUNCTUATION = "DEACTIVATE_PUNCTUATION",
  ACTIVATE_SILENCES = "ACTIVATE_SILENCES",
  DEACTIVATE_SILENCES = "DEACTIVATE_SILENCES",
  ACTIVATE_TEMPORAL_SUGGESTIONS = "ACTIVATE_TEMPORAL_SUGGESTIONS",
  DEACTIVATE_TEMPORAL_SUGGESTIONS = "DEACTIVATE_TEMPORAL_SUGGESTIONS",
  ACTIVATE_PUNCTUATION_SUGGESTIONS = "ACTIVATE_PUNCTUATION_SUGGESTIONS",
  DEACTIVATE_PUNCTUATION_SUGGESTIONS = "DEACTIVATE_PUNCTUATION_SUGGESTIONS",
  ACTIVATE_FORMATTING_SUGGESTIONS = "ACTIVATE_FORMATTING_SUGGESTIONS",
  DEACTIVATE_FORMATTING_SUGGESTIONS = "DEACTIVATE_FORMATTING_SUGGESTIONS",
  ACTIVATE_PASSIVE_ECHO = "ACTIVATE_PASSIVE_ECHO",
  DEACTIVATE_PASSIVE_ECHO = "DEACTIVATE_PASSIVE_ECHO",
  INCREASE_VOLUME = "INCREASE_VOLUME",
  DECREASE_VOLUME = "DECREASE_VOLUME",
  INCREASE_ECHO_RATE = "INCREASE_ECHO_RATE",
  DECREASE_ECHO_RATE = "DECREASE_ECHO_RATE",
  INCREASE_PLAYBACK_RATE = "INCREASE_PLAYBACK_RATE",
  DECREASE_PLAYBACK_RATE = "DECREASE_PLAYBACK_RATE",
  DELETE_SELECTION = "DELETE_SELECTION",
  UPDATE_SELECTION = "UPDATE_SELECTION",
  COPY_SELECTION = "COPY_SELECTION",
  CUT_SELECTION = "CUT_SELECTION",
  EXPORT_SELECTION = "EXPORT_SELECTION",
  RUN_SELECTION = "RUN_SELECTION",
  WALK_SELECTION = "WALK_SELECTION",
  ENTER_SELECTION = "ENTER_SELECTION",
  REMOVE_SELECTION = "REMOVE_SELECTION",
  EXPAND_SELECTION = "EXPAND_SELECTION",
  REDUCE_SELECTION = "REDUCE_SELECTION",
  PLAY_SELECTION = "PLAY_SELECTION",
  ECHO_SELECTION = "ECHO_SELECTION",
  PAUSE_RUN = "PAUSE_RUN",
  EXIT_WALK = "EXIT_WALK",
  SHIFT_ANCHOR_RIGHT = "SHIFT_ANCHOR_RIGHT",
  SHIFT_ANCHOR_LEFT = "SHIFT_ANCHOR_LEFT",
  SHIFT_FOCUS_RIGHT = "SHIFT_FOCUS_RIGHT",
  SHIFT_FOCUS_LEFT = "SHIFT_FOCUS_LEFT",
  SHIFT_SELECTION_FORWARD = "SHIFT_SELECTION_FORWARD",
  SHIFT_SELECTION_BACKWARD = "SHIFT_SELECTION_BACKWARD",
  SHIFT_NEXT_WALK_ELEMENT = "SHIFT_NEXT_WALK_ELEMENT",
  SHIFT_PREVIOUS_WALK_ELEMENT = "SHIFT_PREVIOUS_WALK_ELEMENT",
  SELECT_WORD = "SELECT_WORD",
  SELECT_SENTENCE = "SELECT_SENTENCE",
  SELECT_PARAGRAPH = "SELECT_PARAGRAPH",
  ACCEPT_SELECTION_UPDATE = "ACCEPT_SELECTION_UPDATE",
  REDO_SELECTION_UPDATE = "REDO_SELECTION_UPDATE",
  CANCEL_SELECTION_UPDATE = "CANCEL_SELECTION_UPDATE",
  INCREASE_SELECTION_RATE = "INCREASE_SELECTION_RATE",
  DECREASE_SELECTION_RATE = "DECREASE_SELECTION_RATE",
  PLAY_COMMIT = "PLAY_COMMIT",
  ECHO_COMMIT = "ECHO_COMMIT",
  SELECT_COMMIT = "SELECT_COMMIT",
  ROLLBACK_COMMIT = "ROLLBACK_COMMIT",
  WALK_COMMIT = "WALK_COMMIT",
  RUN_COMMIT = "RUN_COMMIT",
  INSPECT_CLIPBOARD = "INSPECT_CLIPBOARD",
  PASTE_CLIPBOARD = "PASTE_CLIPBOARD",
  ENTER_DICTIONARY = "ENTER_DICTIONARY",
  EXIT_DICTIONARY = "EXIT_DICTIONARY",
  EXPORT_AUDIO = "EXPORT_AUDIO",
  EXPORT_TEXT = "EXPORT_TEXT",
  UNDO_CHANGE = "UNDO_CHANGE",
  REDO_CHANGE = "REDO_CHANGE",
  GRANT_PERMISSION = "GRANT_PERMISSION",
  CANCEL_DIALOG = "CANCEL_DIALOG",
  CONTINUE_DIALOG = "CONTINUE_DIALOG",
}